# Performance improvement after output refactoring