In an attempt to bring our fork closer to master, we discovered that d4c0e814696695bbf853c48b38b7479e0f83f6c7 introduced a meaningful performance change. d4c0e814696695bbf853c48b38b7479e0f83f6c7 replaced several appendStringInfoString uses with appendStringInfo, which dramatically cleaned up the output code at the cost of a performance hit.

This proposal reduces decoding time by ~20% as measured by [this benchmarking script](https://gist.github.com/dpirotte/292ab10510653339dd584413011f1495). It swaps `appendStringInfo` for a new `appendStringInfoStrings` helper, which loops through a variable argument list and calls the more performant `appendStringInfoString` under the hood, skipping if the argument starts with a null byte, i.e. whitespace args when `pretty-print` isn't set. (I think this is safe in these use cases, and it is responsible for a full quarter of the performance return.)

I'm very open to suggestions or other approaches that keep the (very nice) output refactoring and reclaim most of the performance.