@dpirotte 2018-11-07
I ran some additional benchmarks overnight with the same script to verify the performance delta, and it remains consistent.

|pgbench txns|rows output|before|after|diff|
|--------------|-------------|------|-----|---|
|100,000|1,200,057|15.219s|12.058s|-20.8%|
|200,000|1,900,073|23.849s|18.227s|-23.6%|
|300,000|2,600,083|31.222s|24.182s|-22.6%|
|400,000|3,300,101|39.237s|31.146s|-20.7%|
|500,000|4,000,111|47.055s|36.420s|-22.7%|