@jmealo 2019-04-06
@eulerto: this looks interesting. Thoughts?